[dispatch] Add TFP reasoner; add more complex TFP analysis
