+++ analyzer/analysis/tfpAnalysis.py
@@ -10,8 +10,10 @@ from ..shared.utils import *
 from ..shared.logger import *
 from ..shared.config import *
 from ..shared.astTransform import *
+from ..scanner.annotations import *
 from ..analysis.dependencyGraph import DepGraph, is_expr_controlled
-from ..scanner.annotations import get_load_annotation, LoadAnnotation
+from ..scanner.annotations import get_load_annotation
+from .transmissionAnalysis import canonicalize
 # autopep8: on
 
 l = get_logger("TFPAnalysis")
@@ -27,6 +29,7 @@ def get_dependency_graph(t: TaintedFunctionPointer):
     d.resolve_dependencies()
     return d
 
+
 def is_same_var(expr: claripy.ast.BV, reg):
     syms = get_vars(expr)
     assert (len(syms) == 1)
@@ -34,6 +37,7 @@ def is_same_var(expr: claripy.ast.BV, reg):
 
     l.info(f"Testing {sym.args[0]} against {reg}")
     return sym.args[0] == reg
+
 
 def is_potential_secret(d: DepGraph, expr: claripy.ast.BV, tfp_expr: claripy.ast.BV):
     """
@@ -54,7 +58,7 @@ def is_potential_secret(d: DepGraph, expr: claripy.ast.BV, tfp_expr: claripy.ast
     return has_load_anno
 
 
-def analyse(t: TaintedFunctionPointer):
+def analyse(t: TaintedFunctionPointer) -> list[TaintedFunctionPointer]:
     l.warning(f"========= [TFP] ==========")
 
     substitutions = []
@@ -99,36 +103,68 @@ def analyse(t: TaintedFunctionPointer):
     # Analyse tfps
     final_tfps = []
     for tfp in tfps:
-        # If the TFP is not really controlled, skip.
-        if not is_sym_expr(tfp.expr) or not is_expr_controlled(tfp.expr):
-            continue
+
+        if not global_config['NonTaintedFunctionPointers']:
+            # If the TFP is not really controlled, skip.
+            if not is_sym_expr(tfp.expr) or not is_expr_controlled(tfp.expr):
+                continue
 
         d = get_dependency_graph(tfp)
+        tfp_controlled = is_expr_controlled(tfp.expr)
 
         # Analyse registers control.
         for r in tfp.registers:
             if tfp.registers[r].reg == tfp.reg:
-                tfp.registers[r].control = TFPRegisterControlType.IS_TFP_REGISTER
+                tfp.registers[r].control_type = TFPRegisterControlType.IS_TFP_REGISTER
             elif not is_sym_expr(tfp.registers[r].expr) or not is_expr_controlled(tfp.registers[r].expr):
-                tfp.registers[r].control = TFPRegisterControlType.UNCONTROLLED
-                tfp.uncontrolled.append(r)
-            elif not (d.is_independently_controllable(tfp.registers[r].expr, [tfp.expr], check_constraints=True, check_addr=False)
-                      and d.is_independently_controllable(tfp.expr, [tfp.registers[r].expr], check_constraints=True, check_addr=False)):
-                tfp.registers[r].control = TFPRegisterControlType.DEPENDS_ON_TFP_EXPR
+                tfp.registers[r].control_type = TFPRegisterControlType.UNCONTROLLED
+                # Do not add dereferenced registers (not interesting)
+                if not tfp.registers[r].is_dereferenced:
+                    tfp.uncontrolled.append(r)
+            elif tfp_controlled and not (d.is_independently_controllable(tfp.registers[r].expr, [tfp.expr], check_constraints=True, check_addr=False)
+                                         and d.is_independently_controllable(tfp.expr, [tfp.registers[r].expr], check_constraints=True, check_addr=False)):
+                tfp.registers[r].control_type = TFPRegisterControlType.DEPENDS_ON_TFP_EXPR
                 tfp.aliasing.append(r)
-            elif not (d.is_independently_controllable(tfp.registers[r].expr, [tfp.expr], check_constraints=True, check_addr=True)
-                      and d.is_independently_controllable(tfp.expr, [tfp.registers[r].expr], check_constraints=True, check_addr=True)):
-                tfp.registers[r].control = TFPRegisterControlType.INDIRECTLY_DEPENDS_ON_TFP_EXPR
+            elif tfp_controlled and not (d.is_independently_controllable(tfp.registers[r].expr, [tfp.expr], check_constraints=True, check_addr=True)
+                                         and d.is_independently_controllable(tfp.expr, [tfp.registers[r].expr], check_constraints=True, check_addr=True)):
+                tfp.registers[r].control_type = TFPRegisterControlType.INDIRECTLY_DEPENDS_ON_TFP_EXPR
                 tfp.aliasing.append(r)
             elif is_sym_var(tfp.registers[r].expr) and is_same_var(tfp.registers[r].expr, tfp.registers[r].reg):
-                tfp.registers[r].control = TFPRegisterControlType.UNMODIFIED
+                tfp.registers[r].control_type = TFPRegisterControlType.UNMODIFIED
                 tfp.unmodified.append(r)
             elif is_potential_secret(d, tfp.registers[r].expr, tfp.expr):
-                tfp.registers[r].control = TFPRegisterControlType.POTENTIAL_SECRET
+                tfp.registers[r].control_type = TFPRegisterControlType.POTENTIAL_SECRET
                 tfp.secrets.append(r)
             else:
-                tfp.registers[r].control = TFPRegisterControlType.CONTROLLED
-                tfp.controlled.append(r)
+                tfp.registers[r].control_type = TFPRegisterControlType.CONTROLLED
+                # We add controlled registers later
+
+        # Initialize controlled expr for each register
+        for r in tfp.registers:
+            expr = tfp.registers[r].expr
+
+            if expr == None:
+                continue
+
+            try:
+                canonical_exprs = canonicalize(expr, tfp.pc)
+            except SplitTooManyNestedIfException:
+                continue
+
+            controlled_members = []
+            controlled_members = []
+            for canonical_expr in canonical_exprs:
+                members = extract_summed_vals(canonical_expr.expr)
+
+                for ast in members:
+                    if is_attacker_controlled(ast):
+                        controlled_members.append(ast)
+
+            if len(controlled_members) > 0:
+                tfp.registers[r].controlled_expr = generate_addition(
+                    controlled_members)
+            else:
+                tfp.registers[r].controlled_expr = None
 
         final_tfps.append(tfp)
