+++ analyzer/scanner/scanner.py
@@ -35,6 +35,8 @@ from ..asmprinter.asmprinter import get_stack_trace_text
 l = get_logger("Scanner")
 
 n_concrete_addr = 0
+
+
 class DummyConcretizationStrategy(SimConcretizationStrategy):
     """
     Dummy concretization strategy to make ANGR happy. We never use this.
@@ -44,6 +46,7 @@ class DummyConcretizationStrategy(SimConcretizationStrategy):
         global n_concrete_addr
         n_concrete_addr += 8
         return [n_concrete_addr]
+
 
 def skip_concretization(state: angr.SimState):
     """
@@ -62,6 +65,7 @@ class SubstType(Enum):
     CALL_SUBST = 3,
     UNKNOWN_SUBST = 4
 
+
 def getPrefix(stype: SubstType):
     if stype == SubstType.ADDR_SUBST:
         return "addr_subst_"
@@ -73,8 +77,10 @@ def getPrefix(stype: SubstType):
         return "call_subst_"
     return ""
 
+
 def recordSubstitution(state, addr, value, type):
     state.globals[getPrefix(type) + str(addr)] = value
+
 
 def getSubstitution(state, addr, type: SubstType):
     key = getPrefix(type) + str(addr)
@@ -88,6 +94,7 @@ def getSubstitution(state, addr, type: SubstType):
 class SplitException(Exception):
     "Splitted state, skipping"
     pass
+
 
 class Scanner:
     """
@@ -210,6 +217,28 @@ class Scanner:
                 break
         return n_instr
 
+    def count_control_flow_changes(self, state):
+        # Count the number of control flow changes
+        # - All TAKEN Direct/Indirect Calls/Jumps
+        # - Returns
+        n_control_flow_changes = 0
+
+        for bbl_addr, jump_kind, jump_target in zip(self.get_bbls(state)[:-1], list(state.history.jumpkinds)[1:], state.history.jump_targets):
+            if jump_kind == 'Ijk_Ret':
+                n_control_flow_changes += 1
+
+            elif jump_kind == 'Ijk_Call':
+                n_control_flow_changes += 1
+
+            elif jump_kind == 'Ijk_Boring':
+
+                if len(self.bbs[bbl_addr]['block'].vex.constant_jump_targets_and_jumpkinds.keys()) > 1:
+                    # conditional jump, only count if we took non-default target
+                    if self.bbs[bbl_addr]['block'].vex.default_exit_target != jump_target.args[0]:
+                        n_control_flow_changes += 1
+
+        return n_control_flow_changes
+
     def get_aliases(self, state):
         aliases = []
         for v in state.globals.keys():
@@ -278,6 +307,10 @@ class Scanner:
         Indirect calls that are attacker-controlled are saved
         as tainted function pointers (a.k.a Dispatch Gadgets).
         """
+
+        if not global_config['TaintedFunctionPointers']:
+            return
+
         l.warning(
             f"Found new Dispatch Gadget! {func_ptr_ast} {get_annotations(func_ptr_ast)}")
         # Create a new TFP object.
@@ -290,6 +323,8 @@ class Scanner:
                                      constraints=self.get_constraints(state),
                                      n_instr=self.count_instructions(
                                          state, state.scratch.ins_addr),
+                                     n_control_flow_changes=self.count_control_flow_changes(
+                                         state),
                                      contains_spec_stop=self.history_contains_speculation_stop(
                                          state),
                                      n_dependent_loads=get_load_depth(
@@ -299,6 +334,32 @@ class Scanner:
         for reg in get_x86_registers():
             reg_ast = getattr(state.regs, reg)
             tfp.registers[reg] = TFPRegister(reg, reg_ast)
+
+        if global_config['TaintedFunctionPointersRegisterDereference']:
+            # Check if there is a register at dereference
+            for reg in get_x86_registers():
+                reg_expr = tfp.registers[reg].expr
+
+                # If the instruction is a call, it will add 8 to rsp, so we also
+                # need to add + 8
+                call_offset = 0
+                if reg == 'rsp':
+                    block = state.block()
+                    _, regs_write = block.capstone.insns[-1].insn.regs_access()
+                    if 'rsp' in regs_write:
+                        call_offset = 8
+
+                for offset in range(-0x40, 0xa0, 8):
+                    reg_str = f'{reg}_{offset + call_offset}'
+
+                    alias_store, stored_val = memory.get_aliasing_store(
+                        reg_expr + offset + call_offset, 8, self.cur_id, state)
+
+                    if alias_store and is_attacker_controlled(stored_val) and 'gs' not in str(stored_val):
+                        tfp.registers[reg_str] = TFPRegister(
+                            reg_str, stored_val, is_dereferenced=True)
+                        tfp.registers[reg].reg_dereferenced.append(
+                            tfp.registers[reg_str])
 
         self.calls.append(tfp)
