+++ tests/test-cases/.gitignore
@@ -8,4 +8,5 @@ unsupported.txt
 */gadgets.csv
 */gadgets-reasoned.csv
 */tfp.csv
+*/tfp-reasoned.csv
 */half.csv
