+++ tests/test-queries/reasoner-tfp/exploitable.sql
@@ -1,0 +1,9 @@
+SELECT
+    pc,
+    control_sufficiently,
+    control_fully,
+    exploitable,
+    fail_reasons
+FROM
+    tfps
+order by pc, expr, constraints, branches
