+++ tests/test-queries/reasoner-tfp/exploitable_with_branches.sql
@@ -1,0 +1,9 @@
+SELECT
+    pc,
+    control_sufficiently_w_branches,
+    control_fully_w_branches,
+    exploitable_w_branches,
+    fail_reasons_w_branches
+FROM
+    tfps
+order by pc, expr, constraints, branches
